refactor: consolidate unit tests into source files
- Update testing documentation to specify tests should be in `mod tests` blocks
- Move all *_test.rs files into their respective source files:
  - select_test.rs → select.rs
  - attach_test.rs → attach.rs
  - delete_test.rs → delete.rs
  - where_cmd_filesystem_test.rs → where_cmd.rs
- Fix filesystem mock tests to use correct operations (IsDir vs Metadata)
- Remove obsolete test module declarations
- Update CONTRIBUTING.md with clear test organization guidelines

This change improves maintainability by keeping tests close to the code they test
and establishes a consistent testing convention across the codebase.
