+++ rust/src/cli/handlers/where_cmd.rs
@@ -325,4 +325,71 @@ mod tests {
         let result = handle(args, context).await;
         assert!(result.is_ok());
     }
+
+    // Filesystem mocking tests moved from where_cmd_filesystem_test.rs
+    #[tokio::test]
+    async fn test_where_with_filesystem_mocking() {
+        let mut mock = MockCommandExecutor::new();
+        let mock_fs = MockFileSystem::new();
+
+        // Mock git root check
+        mock.expect_command("git").with_args(&["rev-parse", "--git-common-dir"]).returns_output(
+            "/repo/.git",
+            "",
+            0,
+        );
+
+        // Mock filesystem check for worktree existence
+        mock_fs.expect(FileSystemExpectation {
+            operation: FileSystemOperation::IsDir,
+            path: Some(PathBuf::from("/repo/.git/phantom/worktrees/test")),
+            from_path: None,
+            to_path: None,
+            contents: None,
+            result: Ok(MockResult::Bool(true)), // Directory exists
+        });
+
+        let context =
+            HandlerContext::new(mock, mock_fs, crate::core::exit_handler::MockExitHandler::new());
+        let args = WhereArgs { name: Some("test".to_string()), fzf: false, json: false };
+
+        let result = handle(args, context).await;
+        assert!(result.is_ok());
+    }
+
+    #[tokio::test]
+    async fn test_where_worktree_not_found_with_filesystem_mocking() {
+        let mut mock = MockCommandExecutor::new();
+        let mock_fs = MockFileSystem::new();
+
+        // Mock git root check
+        mock.expect_command("git").with_args(&["rev-parse", "--git-common-dir"]).returns_output(
+            "/repo/.git",
+            "",
+            0,
+        );
+
+        // Mock filesystem check - worktree doesn't exist
+        mock_fs.expect(FileSystemExpectation {
+            operation: FileSystemOperation::IsDir,
+            path: Some(PathBuf::from("/repo/.git/phantom/worktrees/nonexistent")),
+            from_path: None,
+            to_path: None,
+            contents: None,
+            result: Ok(MockResult::Bool(false)), // Directory does not exist
+        });
+
+        let context =
+            HandlerContext::new(mock, mock_fs, crate::core::exit_handler::MockExitHandler::new());
+        let args = WhereArgs { name: Some("nonexistent".to_string()), fzf: false, json: false };
+
+        let result = handle(args, context).await;
+        assert!(result.is_err());
+        let error_str = result.unwrap_err().to_string();
+        assert!(
+            error_str.contains("not found") || error_str.contains("Failed to locate"),
+            "Unexpected error message: {}",
+            error_str
+        );
+    }
 }
